Fixes according to the previous comments
